requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
asyncpg==0.29.0
pinecone-client==2.2.4
openai==1.3.7
aiohttp==3.9.1
PyPDF2==3.0.1
python-docx==0.8.11
python-multipart==0.0.6
numpy==1.24.3
